\name{deldata}
\docType{data}
\alias{deldata}
\title{Example Deletion Data}
\description{
This data set gives the deletion data (A/B) from five plant lines for six marker loci. 
}
\usage{deldata}
\format{A matrix containing 5 rows (plants) and 6 columns (marker loci)}
\source{Example dataset}
\keyword{datasets}


