\name{delmap.data}
\alias{delmap.data}
\title{Data Class for Deletion Data
}
\description{
\code{delmap.data} is an delmap object class.
}
\value{
An object of class \code{delmap.data} contains a matrix of 0,1 values where 1's denote deletions. 
It also contains the line names and marker loci names.
}
\author{
Andrew W. George
}
\seealso{
\code{\link{ReadData}}
}


